Add today and work todo lists
- Add navigation for today and work lists
- Add option for removing from lists

+++ index.js
@@ -1,9 +1,11 @@
 import express from "express";
 import bodyParser from "body-parser";
 import mongoose from "mongoose";
+import _ from 'lodash';
 
 const app = express();
 const port = 3000;
+const defaultListName = "Today";
 
 app.use(express.static("public"));
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -32,48 +34,64 @@ const listSchema = {
 
 const List = mongoose.model("List", listSchema);
 
-
 app.get("/", async (req, res) => {
     const tasks = await Task.find({});
     if(tasks.length === 0) {
         await Task.insertMany(defaultTasks);
         res.redirect("/");
     } else {
-        res.render("index.ejs", {taskList: tasks});
+        res.render("index.ejs", {listTitle: defaultListName, taskList: tasks});
     }
 });
 
+app.get("/favicon.ico", (req,res)=>{
+    return "your favicon";
+})   
+
 app.get("/:taskType", async(req, res) => {
-    const taskType = req.params.taskType;
+    const taskType = _.capitalize(req.params.taskType);
     const tasksList = await List.findOne({name: taskType});
     if(!tasksList){
         const list = new List({
             name: taskType,
             tasks: defaultTasks
         })
-        list.save();
-        console.log('doesnt exist');
+        await list.save();
+        res.redirect("/" + taskType);
     } else {
-        res.render("index.ejs", {listTitle: tasksList.name, newTasks: tasksList.tasks});
+        res.render("index.ejs", {listTitle: tasksList.name, taskList: tasksList.tasks});
     }
   })
 
 app.post("/", async(req, res) => {
     const taskName = req.body.task;
+    const listName = req.body.listTitle;
     const task = new Task({
         name: taskName
     });
-    task.save();
 
-    res.redirect("/");
+    if(listName === defaultListName) {
+        task.save();
+        res.redirect("/");
+    } else {
+        const foundList = await List.findOne({name: listName});
+        foundList.tasks.push(task);
+        foundList.save();
+        res.redirect("/" + listName);
+    }
 });
 
 app.post("/delete", async(req, res) => {
     const checkedTaskId = req.body.checkbox;
+    const listName = req.body.listName;
 
-    await Task.findByIdAndRemove(checkedTaskId);
-
-    res.redirect("/");
+    if (listName === defaultListName) {
+        await Task.findByIdAndRemove(checkedTaskId);
+        res.redirect("/");
+    } else {
+        await List.findOneAndUpdate({name: listName}, { $pull: {tasks: {_id: checkedTaskId}}});
+        res.redirect("/" + listName);
+    }
 });
 
 app.listen(port, () => {
